backend/requirements.txt
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
numpy==1.26.4
scikit-learn==1.3.2
scipy==1.12.0 